print-spike-1.0.0 Updates

+++ release.sh
@@ -10,22 +10,26 @@ cp address.erb ../release/address-service/$ADDRESS_SERVICE.erb
 
 
 # create tar for olcs-addressbase (address ETL scripts)
-rm -rf olcs-addressbase
-git clone git@gitlab.inf.mgt.mtpdvsa:sc/address-base.git olcs-addressbase
-cd olcs-addressbase
-git checkout $ADDRESS_ETL
-tar -czvf ../../release/olcs-addressbase/$ADDRESS_ETL.tar.gz *
-cd ..
+if [ ! -f ../../release/olcs-addressbase/$ADDRESS_ETL.tar.gz ]; then
+    rm -rf olcs-addressbase
+    git clone git@gitlab.inf.mgt.mtpdvsa:sc/address-base.git olcs-addressbase
+    cd olcs-addressbase
+    git checkout $ADDRESS_ETL
+    tar -czvf ../../release/olcs-addressbase/$ADDRESS_ETL.tar.gz *
+    cd ..
+fi
 
 # create tar for olcs-elasticsearch scripts
-rm -rf olcs-elasticsearch
-git clone git@gitlab.inf.mgt.mtpdvsa:olcs/olcs-elasticsearch.git
-cd olcs-elasticsearch
-git checkout $ELASTIC
-tar -czvf ../../release/olcs-elasticsearch/$ELASTIC.tar.gz *
-cd ..
+if [ ! -f ../../release/olcs-elasticsearch/$ELASTIC.tar.gz ]; then
+    rm -rf olcs-elasticsearch
+    git clone git@gitlab.inf.mgt.mtpdvsa:olcs/olcs-elasticsearch.git
+    cd olcs-elasticsearch
+    git checkout $ELASTIC
+    tar -czvf ../../release/olcs-elasticsearch/$ELASTIC.tar.gz *
+    cd ..
+fi
 
-# OpenAM NB This openAM bit hasn;t been tested Mat 26 Feb 2016
+# OpenAM
 rm -rf olcs-oa
 git clone git@gitlab.inf.mgt.mtpdvsa:olcs/olcs-oa.git
 cd olcs-oa
@@ -37,19 +41,22 @@ cd ..
 
 
 # create tar for olcs-templates
-rm -rf olcs-templates
-git clone git@gitlab.inf.mgt.mtpdvsa:olcs/olcs-templates.git
-cd olcs-templates
-git checkout $TEMPLATES
-tar -czvf ../../release/olcs-templates/$TEMPLATES.tar.gz *
-cd ..
-
-
-if [ ! -d olcs-txt ]; then
-    git clone git@gitlab.inf.mgt.mtpdvsa:olcs/olcs-txc.git
+if [ ! -f ../../release/olcs-templates/$TEMPLATES.tar.gz ]; then
+    rm -rf olcs-templates
+    git clone git@gitlab.inf.mgt.mtpdvsa:olcs/olcs-templates.git
+    cd olcs-templates
+    git checkout $TEMPLATES
+    tar -czvf ../../release/olcs-templates/$TEMPLATES.tar.gz *
+    cd ..
 fi
-(cd olcs-txc && git checkout $TXCHANGE)
-cp olcs-txc/txc.war ../release/txc/$TXCHANGE.war
+
+if [ ! -f ../release/txc/$TXCHANGE.war ]; then
+    if [ ! -d olcs-txt ]; then
+        git clone git@gitlab.inf.mgt.mtpdvsa:olcs/olcs-txc.git
+    fi
+    (cd olcs-txc && git checkout $TXCHANGE)
+    cp olcs-txc/txc.war ../release/txc/$TXCHANGE.war
+fi
 
 cd ../release
